arm64/dts: zuma: Add AoC S2MPU
AoC is only controlled by TZ.

However, SysMMU has an emulation feature that can be misused to read
from arbitrary memory locations, and with SysMMU under the control of
the kernel, we need to configure S2MPU to block such potentially
malicious transactions.

Add the AoC S2MPU with the new flag “deny-all” which would mainly
unmap the S2MPU interface and configure it to deny all traffic.

Bug: 204991862
Bug: 342092765
Change-Id: Iaa4aed25fc6b9e09f8a06c92d2fcf3fcc658a7bd
Signed-off-by: Mostafa Saleh <smostafa@google.com>