arm64/dts: zuma: Add alive S2MPU
Alive S2MPU is fully trusted and is not managed by the kernel.
However, we need to make sure the kernel can’t map it to use
it in emulation mode of the S2MPU.

So we define it as:
- “deny-all”: So it never gets power managed or updated
- “off-at-boot”: Never configured at boot.

Which mainly means it remains unconfigured but it's MMIO is not
accessible to the kernel.

Bug: 342511931
Change-Id: Ia027294962b41f19cf31c83fcdf51c3c5b4866cc
Merged-In: I82d45ceb2ba7cf4aad13d8e038e0bead23c582d2
Signed-off-by: Mostafa Saleh <smostafa@google.com>