vh: sched: Add a separate disable_util_est knob
Instead of overloading the meaning of rampup_multiplier to imply
disabling util_est when it is 0, add an explicit knob to set that.

We could have situations where we want the multiplier to be 0, but
util_est is not unconditionally 0 too. This new knob should allow us to
control the behavior per group.

This knob is only applied when auto_dvfs_headroom is enabled.

Only honour vp->ignore_util_est_update if rampup_multiplier is not 0.

Bug: 336169796
Signed-off-by: Qais Yousef <qyousef@google.com>
Change-Id: I24f6072a98461881b440870182dd744c1ad92eff