usb: xhci-exynos: support suspend for accessory under a hub
Support auto-suspend when the audio accessory is connected to the
device through a hub. When a hub is connected, we will scan all ports
of the hub. The System is allowed to suspend only when all the
connected accessories support remote wakeup.

ported from I7db3d7ebab339b4b86ec7ac600cc10db044ae68d

Bug: 353613572
Change-Id: I8bdf9d89c705721adccbfaeb8cfb0bbe28aa3ba0
Signed-off-by: Kyle Tso <kyletso@google.com>