xhci-exynos: Maintain a list of udev requiring autosuspend
xhci-exynos requires the udev to set the remote_wakeup bit in the
descriptor to enable autosuspend feature. However, some accessories do
not report remote_wakeup capability to the Host and still work well if
the Host suspends the bus. Maintain a list of the VID:PID pair of such
accessories so that xhci-exynos is able to autosuspend when these
accessories are enumerated.

ported from Ief3e67a19471d8cc730e2792ea44797c76957c4a

Bug: 353613572
Change-Id: I921db13ff7cfaa5d8cbf77782d14f95df0b657d5
Signed-off-by: Kyle Tso <kyletso@google.com>