pcie: exynos: Avoid blocking on cfg_access work while atomic
Config register reads and IRQ handlers are both atomic contexts -- the
former because all config registers hold a spin lock (with
CONFIG_PCI_LOCKLESS_CONFIG=n), and the latter because it's not a
threaded interrupt handler.

In atomic contexts, we shouldn't make blocking (sleeping) calls, as they
may simply hang.

Per my reasoning on $BUG, we're in one of the following cases:
1. the work item hasn't run yet (we haven't timed out yet); cancel_*()
   is equivalent to cancel_*sync();
2. the work will do nothing (since we're in cpl_timeout_recovery), and
   therefore there's nothing to wait and race for -- it's OK to let the
   work-item finish running on its own schedule; or
3. we've truly timed out, and the work item will run its dump-and-crash
   work.

Option 3 is the most critical case, but:
(a) in such cases, we expect to hang in the read anyway, so we're not
    actually going to race; and
(b) even if somehow the read is unblocked, it's still OK to be
    dumping in parallel. We were going to BUG_ON(1) and panic at the end
    of the dump anyway.

Bug: 359291551
Test: build --config=debug_all; boot; fewer sched-debug warnings
Change-Id: Ide20e0c0b1ac81218d48dfde20e7b3024f553fbb
Signed-off-by: Brian Norris <briannorris@google.com>