vh: sched: Use a small margin to detect periodic vs transient
util_avg can have small variations due to the 1024us period and the fact
that util_invariance could be inaccurate if a freq changes for a RUNNING
task.

Add a small margin when comparing task_util() with util_enqueued to
detect if the task is going through a transition.

Addressing the source of these variations instead of using margins is
ideal in the long run.

Bug: 336169796
Signed-off-by: Qais Yousef <qyousef@google.com>
Change-Id: Icd84aeeef18d18184a6da2a0b1f1b9aae1a6e1d9